requirements.txt
# Core validation
jsonschema>=4.20.0

# LLM & Web fetching for paper extraction
dedalus-labs>=0.1.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# CLI improvements (optional but nice)
rich>=13.7.0
click>=8.1.0
